fix: 对金额等类型修改为bigdecimal，相关操作修改为原子性

+++ handler/shouye.go
@@ -259,7 +259,7 @@ func orderToDTO(o model.OrderTbl) OrderDTO {
 		LeaveTime:            timeToString(o.LeaveTime),
 		ChargeGunPull:        o.ChargeGunPull,
 		GunDisconnectTime:    timeToLocal(o.GunDisconnectTime),
-		DecideAmount:         int32(o.DecideAmount),
+		DecideAmount:         o.DecideAmount,
 		NeighborRelocateTime: timeToLocal(o.NeighborRelocateTime),
 		MoveCarTime:          timeToLocal(o.MoveCarTime),
 		ImageID:              o.ImageID,
